Move async task helper functions to util file
